refactor: change archive names and functions
