+++ engineering/admin/src/app.module.ts
@@ -1,7 +1,7 @@
 import { ClassSerializerInterceptor, Module } from '@nestjs/common';
 import { APP_FILTER, APP_INTERCEPTOR } from '@nestjs/core';
-import { FiltroDeExcecaoHttp } from './common/filtros/filtro-de-excexao-http.filter';
-import { EmailBoxModule } from './emailBox/usuario.module';
+import { FiltroDeExcecaoHttp } from './core/common/filtros/filtro-de-excexao-http.filter';
+import { EmailBoxModule } from './emailBox/emailBox.module';
 import { TransformaRespostaInterceptor } from './usuario/core/http/transforma-resposta-interceptor';
 import { UsuarioModule } from './usuario/usuario.module';
